% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OO_classes.R
\docType{class}
\name{CES-class}
\alias{CES-class}
\title{Class definition for pfdecomp CES object}
\description{
Class definition for pfdecomp CES object
}
\section{Slots}{

\describe{
\item{\code{output}}{ts.}

\item{\code{capital}}{ts.}

\item{\code{labour}}{ts.}

\item{\code{beta}}{numeric.}

\item{\code{sigma}}{numeric.}

\item{\code{solow}}{ts.}

\item{\code{contribs}}{mts.}

\item{\code{avg_gr}}{matrix.}
}}

