% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobb_douglas.R
\name{cobb_douglas}
\alias{cobb_douglas}
\title{Growth accounting using Cobb-Douglas production function}
\usage{
cobb_douglas(y, k, l, beta, figure = FALSE, gr_opt = "log")
}
\arguments{
\item{y}{A ts object.}

\item{k}{A ts object.}

\item{l}{A ts object.}

\item{beta}{A numeric value.}

\item{figure}{A boolean value.}

\item{gr_opt}{A character value.}
}
\value{
Cobb-Douglas object.
}
\description{
Growth accounting using Cobb-Douglas production function
}
