% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces.R
\name{ces}
\alias{ces}
\title{Growth accounting using CES production function (with labour augmenting technical progress)}
\usage{
ces(y, k, l, beta, sigma, figure = FALSE, gr_opt = "log")
}
\arguments{
\item{y}{A ts object.}

\item{k}{A ts object.}

\item{l}{A ts object.}

\item{beta}{A numeric value.}

\item{sigma}{A numeric value.}

\item{figure}{A boolean value.}

\item{gr_opt}{A character value.}
}
\value{
CES object.
}
\description{
Growth accounting using CES production function (with labour augmenting technical progress)
}
