% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfdecomp-package.R
\docType{package}
\name{pfdecomp-package}
\alias{pfdecomp}
\alias{pfdecomp-package}
\title{pfdecomp: Production Function Decomposition}
\description{
A package for decomposing Cobb-Douglas and CES production functions into contributions from factors inputs.
}
\author{
\strong{Maintainer}: Graeme Walsh \email{graeme.walsh@hotmail.co.uk} (\href{https://orcid.org/0009-0005-9190-376X}{ORCID})

}
\keyword{internal}
